simulator_mavlink: switch battery max and min to fix simultion
In 4ccb0b3ce3372db2ea54d725dacdb91642e104a7 the simulated battery
percentage was reversed. I'm assuming because of the possibly
missleading variable name. Now I'm fixing it by switching the
maximum and minimum voltage such that the name is not misleading anymore
but it still works as expected.