simulator_mavlink: take care of battery percentage rounding errors
In 42b2d140cc175f895e0f2735e1237a538dbde838 I made MAVLink send a
rounded up integer instead of rounded down. This makes sense in practice
because the low battery reactions happen exactly when the reported number
switches in the UI. But here we want to provoke an exact 50% in the UI so
we stop counting at 49.9%, it get's rounded up and we see the expected
result.