helper_functions: include defines for M_PI_PRECISE
The defintion of the custom symbol M_PI_PRECISE
was not included in one of the only places it's used.
Looks like a mistake that happened in
dbb311c70c0875a74f2c51e719ea403683a48df6
possibly because a lot of things are included
almost everywhere and if the include order
ligns up there are no compile errors.